fix demos

+++ ChartsDemo/Classes/RealmBase/Score.h
@@ -1,6 +1,8 @@
 //
 //  Score.h
 //  ChartsDemo
+//
+//  Created by Daniel Cohen Gindi on 17/01/2015.
 //  Copyright © 2015 dcg. All rights reserved.
 //
 
@@ -9,11 +11,11 @@
 @interface Score : RLMObject
 
 @property (nonatomic, assign) float totalScore;
-@property (nonatomic, assign) double scoreNr;
+@property (nonatomic, assign) NSInteger scoreNr;
 @property (nonatomic, strong) NSString *playerName;
 
 - (id)initWithTotalScore:(float)totalScore
-                 scoreNr:(double)scoreNr
+                 scoreNr:(NSInteger)scoreNr
               playerName:(NSString *)playerName;
 
 @end
