+++ ChartsDemo/Classes/RealmBase/Score.m
@@ -1,6 +1,8 @@
 //
 //  Score.m
 //  ChartsDemo
+//
+//  Created by Daniel Cohen Gindi on 17/01/2015.
 //  Copyright © 2015 dcg. All rights reserved.
 //
 
@@ -9,7 +11,7 @@
 @implementation Score
 
 - (id)initWithTotalScore:(float)totalScore
-                 scoreNr:(double)scoreNr
+                 scoreNr:(NSInteger)scoreNr
               playerName:(NSString *)playerName
 {
     self = [super init];
